Update news card and page
