+++ l10n/intl_en.arb
@@ -165,7 +165,7 @@
   "ets_gus": "GUS",
   "ets_papercut_title":"PaperCut",
 
-  "news_title" : "News",
+  "news_title" : "Announcement",
 
   "more_about_applets_title": "About ApplETS",
   "more_report_bug": "Report a bug or request a feature",
