+++ l10n/intl_fr.arb
@@ -165,7 +165,7 @@
   "ets_gus": "GUS",
   "ets_papercut_title":"PaperCut",
 
-  "news_title" : "Nouvelles",
+  "news_title" : "Annonces",
 
   "more_about_applets_title": "À propos d'ApplETS",
   "more_report_bug": "Rapporter un bogue ou une amélioration",
